requirements.txt
langchain>=0.1.0
langchain-community>=0.0.1
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.1.1
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
